feat: rewrite on ts
